refactor: Post 테이블 컴포넌트(PostTable) 페이지에서 분리
